Added changes for latest Kivio from theKompany's CVS.  This however, overwrote the kdDebug changes you guys made.  Mine are almost the same, but don't end in a '\n'.  I didn't know you were supposed to do that, so kivio now generates a ton of ASSERT failures because of this.  But I kept all the other changes such as the shell removal (I think), and the i18n modifications.  All changes I made locally to the source were also made.  Don't do any source modifications yet.
svn path=/trunk/koffice/; revision=74890

+++ kivio/kiviopart/kivio_stencil_geometry_panel.h
@@ -12,21 +12,21 @@ protected:
 
 public slots:
     void setUnit(int);
-    void setPosition( double, double );
-    void setSize( double, double );
+    void setPosition(float, float);
+    void setSize(float, float);
 
 protected slots:
-    void xChange( double );
-    void yChange( double );
-    void wChange( double );
-    void hChange( double );
+    void xChange(float);
+    void yChange(float);
+    void wChange(float);
+    void hChange(float);
 
 signals:
-    void positionChanged( double, double );
-    void sizeChanged( double, double );
+    void positionChanged(float, float);
+    void sizeChanged(float, float);
 
 public:
-    KivioStencilGeometryPanel( QWidget *parent );
+    KivioStencilGeometryPanel(QWidget* parent);
     virtual ~KivioStencilGeometryPanel();
 };
